# asset-service/docs/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Asset Management Service API
  description: |
    Asset Management Service API
    
    Comprehensive asset lifecycle management including:
    - Asset CRUD operations
    - Master data management (Categories, Makes, Models, Vendors, etc.)
    - User-Asset linking
    - Compliance and validation
    - Document management
    - Warranty and AMC management
    - Audit logging
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:8082
    description: Local development server
  - url: https://api.example.com/asset-service
    description: Production server
...
paths:
...
  /api/asset/v1/compliance/status/{entityType}/{entityId}:
    get:
      tags:
        - Compliance
      summary: Get compliance status
      description: Get the compliance status of an entity
      security:
        - bearerAuth: []
      parameters:
        - name: entityType
          in: path
          required: true
          schema:
            type: string
        - name: entityId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Compliance status retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  violations:
                    type: array
                    items:
                      type: object
  /api/asset/v1/compliance/violations/{entityType}/{entityId}:
    get:
      tags:
        - Compliance
      summary: Get compliance violations
      description: Get all compliance violations for an entity
      security:
        - bearerAuth: []
      parameters:
        - name: entityType
          in: path
          required: true
          schema:
            type: string
        - name: entityId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Violations retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...